Added file cleanup to reduction loops
To resolve some weird testing errors for the various reduction methods,
I had to make the reduction functions include number of species in
the filename (to avoid conflicts). But, then the reductions generate a large
number of files. Now, they remove the file during the iterations, and only
save the file at the end of the loop.