Changes most reduction inputs to YAML file. Closes #59, closes #58, and closes #57.
These changes move all reduction-specific inputs to a YAML file, which
also includes the sampling input parameters (autoignition only for now).
As part of this, pyMARS now checks the inputs for correctness and consistency,
including that species included as targets, always retained, and as
fuel/oxidizer/reactant are actually in the starting model. Tests are
updated to reflect these changes, though documentation has not yet
been updated.