Added laminar flame speed simulations as an option to base reduction on. This involved some light renaming of variables in the code to be more generalizable. For the laminar flame speed simulations, only samples from Tu + 0.05*(Tb - Tu) < T < Tb - 0.05*(Tb - Tu) are collected for DRG analysis, following Ho, J.Z., Talei, M., Gordon, R.L. and Brear, M.J., 2021. Reduced chemistry for sound generation by planar annihilation in premixed methane/hydrogen flames. Proceedings of the Combustion Institute, 38(4), pp.6125-6133. Note that the R_AB weightage depending on the fuel content described in that paper has not yet been implemented.
Also, modified the sensitivity analysis so that it continues running even if it encounters a Cantera error after removing a species. Useful when it removes a species and Cantera crashes, but other species can be removed without issue.

I've only tested this on DRGEP + SA, so it's a very developmental branch. The /tests folder is not updated.